+++ lib/src/reflection_factory_builder.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:collection';
 
 import 'package:analyzer/dart/analysis/results.dart';
@@ -17,13 +18,35 @@ import 'analyzer/library.dart';
 import 'analyzer/type_checker.dart';
 import 'reflection_factory_annotation.dart';
 import 'reflection_factory_base.dart';
+import 'reflection_factory_extension.dart';
 
 /// The reflection builder.
 class ReflectionBuilder implements Builder {
-  /// If `true` builds code in verbose mode.
-  bool verbose;
+  /// If `true` builds the reflection code in verbose mode.
+  final bool verbose;
 
-  ReflectionBuilder({this.verbose = false});
+  /// If `true` will process the [BuildStep]s sequentially. See [build].
+  final bool sequential;
+
+  /// The sequential [BuildStep] timeout (default: 30 sec). See [build].
+  final Duration buildStepTimeout;
+
+  ReflectionBuilder({
+    this.verbose = false,
+    bool sequential = true,
+    this.buildStepTimeout = const Duration(seconds: 30),
+  }) : sequential = sequential && buildStepTimeout.inMilliseconds > 0;
+
+  /// The [ReflectionFactory.VERSION].
+  String get version => ReflectionFactory.VERSION;
+
+  @override
+  String toString({String indent = ''}) {
+    return '${indent}ReflectionBuilder[$version]:\n'
+        '$indent  -- verbose: $verbose\n'
+        '$indent  -- sequential: $sequential\n'
+        '$indent  -- buildStepTimeout: ${buildStepTimeout.toHumanReadable()}\n';
+  }
 
   @override
   final buildExtensions = const {
@@ -41,23 +64,111 @@ class ReflectionBuilder implements Builder {
 
   static const TypeChecker typeClassProxy = TypeChecker.fromRuntime(ClassProxy);
 
+  Future<void> _buildChaing = Future<void>.value();
+
+  /// If [sequential] is `true` every [BuildStep] is processed sequentially (only one at a time):
+  /// - The default timeout to process a [BuildStep] is 30 sec ([buildStepTimeout]).
+  ///   If the timeout is reached the next [BuildStep] starts to be processed.
+  /// - `build_runner` will wait for ALL the [BuildStep]s to complete (regardless of any timeout).
   @override
-  Future<void> build(BuildStep buildStep) async {
-    var inputLib = await buildStep.inputLibrary;
+  Future<void> build(BuildStep buildStep) {
+    // Skip code from `reflection_factory` library.
+    if (_isReflectionFactoryLibraryCode(buildStep.inputId)) {
+      return Future<void>.value();
+    }
+
+    if (!sequential) {
+      return _buildImpl(buildStep);
+    }
+
+    var buildChaing = _buildChaing;
+
+    var buildTimeouted = Completer<void>();
+    void complete(_) => buildTimeouted.complete();
+
+    Future<void> callBuildImpl(_) {
+      var future = _buildImpl(buildStep);
+
+      future
+          .timeout(buildStepTimeout,
+              onTimeout: () => _onBuildTimeout(buildStep))
+          .then(complete, onError: complete);
+
+      return future;
+    }
+
+    var build = buildChaing.then(callBuildImpl, onError: callBuildImpl);
+
+    // The chain wil have a schedulled timeout only
+    // for the last running [BuildStep].
+    _buildChaing = buildTimeouted.future;
+
+    // Return the call without timeout for the `build_runner`.
+    return build;
+  }
+
+  bool _isReflectionFactoryLibraryCode(AssetId assetId) {
+    if (assetId.package == 'reflection_factory') {
+      if (!assetId.path.startsWith('example/') &&
+          !assetId.path.startsWith('test/')) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  void _onBuildTimeout(BuildStep buildStep) {
+    log.warning(
+        'Build timeout (${buildStepTimeout.toHumanReadable()}): ${buildStep.inputId} >> Will continue with the next `BuildStep`...');
+  }
+
+  Future<void> _buildImpl(BuildStep buildStep) async {
+    final initTime = DateTime.now();
+
+    log.info(" analyzing...");
+
+    var resolver = buildStep.resolver;
+
     var inputId = buildStep.inputId;
 
-    if (inputId.package == 'reflection_factory') {
-      if (!inputId.path.startsWith('example/') &&
-          !inputId.path.startsWith('test/')) {
-        return;
-      }
-    } else if (inputLib.name == 'reflection_factory' ||
+    var inputLib = await buildStep.inputLibrary;
+
+    if (inputLib.name == 'reflection_factory' ||
         inputLib.name.startsWith('reflection_factory.')) {
       return;
     }
 
     var libraryReader = LibraryReader(inputLib);
 
+    var genPart = await buildStep.trackStage('Resolving `part` to generate',
+        () => _resolveGeneratedPart(buildStep, inputId, libraryReader));
+
+    if (genPart == null) {
+      final elapsedTime = DateTime.now().difference(initTime);
+      log.info(
+          " no reflection to generate. (${elapsedTime.inMilliseconds} ms)");
+      return;
+    }
+
+    var typeAliasTable = _TypeAliasTable.fromLibraryReader(libraryReader);
+
+    var codeTable = await buildStep.trackStage('Generating reflection code',
+        () => _buildCodeTable(buildStep, libraryReader, typeAliasTable));
+
+    if (codeTable.isEmpty) {
+      throw StateError("Generated code expected for: $inputId");
+    }
+
+    await buildStep.trackStage('Writing generated code',
+        () => _writeFullCode(buildStep, genPart, codeTable, initTime));
+
+    if (resolver is ReleasableResolver) {
+      resolver.release();
+    }
+  }
+
+  Future<_GeneratedPart?> _resolveGeneratedPart(
+      BuildStep buildStep, AssetId inputId, LibraryReader libraryReader) async {
     var inputFile = inputId.pathSegments.last;
 
     var allParts = _readReflectionPartDirectives(libraryReader, inputFile);
@@ -74,12 +185,14 @@ class ReflectionBuilder implements Builder {
 
     var hasCodeToGenerate = await _hasCodeToGenerate(buildStep, libraryReader);
     if (!hasCodeToGenerate) {
-      return;
+      return null;
     }
 
     var genSiblingId = inputId.changeExtension('.reflection.g.dart');
     var genSubId =
         inputId.changeExtension('.g.dart').withParentDirectory('reflection');
+
+    var genId = isSiblingPart ? genSiblingId : genSubId;
 
     // No `part` directive!
     if (siblingParts.isEmpty && subParts.isEmpty) {
@@ -98,20 +211,12 @@ class ReflectionBuilder implements Builder {
           "  > Found part directives: $gParts");
     }
 
-    var typeAliasTable = _TypeAliasTable.fromLibraryReader(libraryReader);
+    return _GeneratedPart(
+        genId.path, genId, isSiblingPart ? inputFile : '../$inputFile');
+  }
 
-    var codeTable =
-        await _buildCodeTable(buildStep, libraryReader, typeAliasTable);
-
-    if (codeTable.isEmpty) {
-      throw StateError("Generated code expected for: $inputId");
-    }
-
-    var genId = isSiblingPart ? genSiblingId : genSubId;
-
-    var reflectionMixin = _buildReflectionMixin(codeTable, typeAliasTable);
-    var siblingsClassReflection = _buildSiblingsClassReflection(codeTable);
-
+  Future<void> _writeFullCode(BuildStep buildStep, _GeneratedPart genPart,
+      _CodeTable codeTable, DateTime initTime) async {
     var fullCode = StringBuffer();
 
     fullCode.write('// \n');
@@ -127,14 +232,10 @@ class ReflectionBuilder implements Builder {
     fullCode.write('// ignore_for_file: unnecessary_cast\n');
     fullCode.write('// ignore_for_file: unnecessary_type_check\n\n');
 
-    if (isSiblingPart) {
-      fullCode.write("part of '$inputFile';\n\n");
-    } else {
-      fullCode.write("part of '../$inputFile';\n\n");
-    }
+    fullCode.write("part of '${genPart.partOf}';\n\n");
 
-    fullCode.write(typeAliasTable.code);
-    fullCode.write(reflectionMixin);
+    fullCode.write(codeTable.typeAliasTable.code);
+    fullCode.write(codeTable.codeReflectionMixin);
 
     var codeKeys = codeTable.allKeys.toList();
     _sortCodeKeys(codeKeys);
@@ -146,21 +247,26 @@ class ReflectionBuilder implements Builder {
       }
     }
 
-    fullCode.write(siblingsClassReflection);
+    fullCode.write(codeTable.codeSiblingsClassReflection);
 
     var generatedCode = fullCode.toString();
 
     var dartFormatter = DartFormatter();
     var formattedCode = dartFormatter.format(generatedCode);
 
+    var genId = genPart.genId;
+
     await buildStep.writeAsString(genId, formattedCode);
 
-    print('** GENERATED:\n'
+    final elapsedTime = DateTime.now().difference(initTime);
+
+    log.info(' >> GENERATED:\n'
         '  -- Elements: $codeKeys\n'
-        '  -- Code file: $genId\n');
+        '  -- File: $genId\n'
+        '  -- Time: ${elapsedTime.inMilliseconds} ms\n\n');
 
     if (verbose) {
-      print('<<<\n$formattedCode\n>>>');
+      log.info('<<<<<<\n$formattedCode\n>>>>>>\n');
     }
   }
 
@@ -244,7 +350,7 @@ class ReflectionBuilder implements Builder {
 
   Future<_CodeTable> _buildCodeTable(BuildStep buildStep,
       LibraryReader libraryReader, _TypeAliasTable typeAliasTable) async {
-    var codeTable = _CodeTable();
+    var codeTable = _CodeTable(typeAliasTable);
 
     var annotatedClassProxy =
         libraryReader.annotatedWith(typeClassProxy).toList();
@@ -294,6 +400,9 @@ class ReflectionBuilder implements Builder {
       }
     }
 
+    _buildReflectionMixin(codeTable);
+    _buildSiblingsClassReflection(codeTable);
+
     return codeTable;
   }
 
@@ -334,12 +443,12 @@ class ReflectionBuilder implements Builder {
       reflectionProxyName = annotatedClass.name;
     }
 
-    print('** ClassProxy:\n'
+    log.info(' <ClassProxy>\n'
         '  -- Target Class Name: $className\n'
         '  -- Always Return Future: $alwaysReturnFuture\n'
         '  -- reflectionProxyName: $reflectionProxyName\n'
         '  -- traverseReturnTypes: $traverseReturnTypes\n'
-        '  -- ignoreParametersTypes: $ignoreParametersTypes\n');
+        '  -- ignoreParametersTypes: $ignoreParametersTypes\n\n');
 
     var codeTable = <String, String>{};
 
@@ -367,7 +476,7 @@ class ReflectionBuilder implements Builder {
     );
 
     if (verbose) {
-      print(classTree);
+      log.info(' >> $classTree');
     }
 
     codeTable.putIfAbsent(
@@ -454,11 +563,11 @@ class ReflectionBuilder implements Builder {
         .mapValue
         .map((k, v) => MapEntry(k!.toTypeValue()!, v!.toStringValue()!));
 
-    print('** ReflectionBridge:\n'
+    log.info(' <ReflectionBridge>\n'
         '  -- classesTypes: $classesTypes\n'
         '  -- bridgeExtensionName: $bridgeExtensionName\n'
         '  -- reflectionClassNames: $reflectionClassNames\n'
-        '  -- reflectionExtensionNames: $reflectionExtensionNames\n');
+        '  -- reflectionExtensionNames: $reflectionExtensionNames\n\n');
 
     var codeTable = <String, String>{};
 
@@ -484,7 +593,7 @@ class ReflectionBuilder implements Builder {
       );
 
       if (verbose) {
-        print(classTree);
+        log.info(' >> $classTree');
       }
 
       codeTable.putIfAbsent(classTree.classGlobalFunction('_'),
@@ -561,7 +670,7 @@ class ReflectionBuilder implements Builder {
     );
 
     if (verbose) {
-      print(enumTree);
+      log.info(' >> $enumTree');
     }
 
     var enumGlobalFunctions = enumTree.buildEnumGlobalFunctions();
@@ -594,7 +703,7 @@ class ReflectionBuilder implements Builder {
     );
 
     if (verbose) {
-      print(classTree);
+      log.info(' >> $classTree');
     }
 
     var classGlobalFunctions = classTree.buildClassGlobalFunctions();
@@ -616,13 +725,12 @@ class ReflectionBuilder implements Builder {
     return library;
   }
 
-  String _buildReflectionMixin(
-      _CodeTable codeTable, _TypeAliasTable typeAliasTable) {
-    if (codeTable.reflectionClassesIsEmpty) return '';
+  void _buildReflectionMixin(_CodeTable codeTable) {
+    if (codeTable.reflectionClassesIsEmpty) return;
 
     var str = StringBuffer();
 
-    str.write('mixin ${typeAliasTable.reflectionMixinName} {\n');
+    str.write('mixin ${codeTable.typeAliasTable.reflectionMixinName} {\n');
 
     str.write(
         "  static final Version _version = Version.parse('${ReflectionFactory.VERSION}');\n\n");
@@ -634,12 +742,11 @@ class ReflectionBuilder implements Builder {
 
     str.write('}\n\n');
 
-    var code = str.toString();
-    return code;
+    codeTable.codeReflectionMixin = str.toString();
   }
 
-  String _buildSiblingsClassReflection(_CodeTable codeTable) {
-    if (codeTable.reflectionClassesIsEmpty) return '';
+  void _buildSiblingsClassReflection(_CodeTable codeTable) {
+    if (codeTable.reflectionClassesIsEmpty) return;
 
     var str = StringBuffer();
 
@@ -667,9 +774,16 @@ class ReflectionBuilder implements Builder {
     str.write('  assert(length > 0);\n');
     str.write('}\n\n');
 
-    var code = str.toString();
-    return code;
+    codeTable.codeSiblingsClassReflection = str.toString();
   }
+}
+
+class _GeneratedPart {
+  final String partPath;
+  final AssetId genId;
+  final String partOf;
+
+  _GeneratedPart(this.partPath, this.genId, this.partOf);
 }
 
 extension _AssetIdExtension on AssetId {
@@ -784,6 +898,13 @@ class _TypeAliasTable {
 }
 
 class _CodeTable {
+  final _TypeAliasTable typeAliasTable;
+
+  String codeReflectionMixin = '';
+  String codeSiblingsClassReflection = '';
+
+  _CodeTable(this.typeAliasTable);
+
   final Map<String, String> _reflectionClasses = <String, String>{};
   final Map<String, String> _reflectionProxies = <String, String>{};
 
@@ -1457,7 +1578,7 @@ class _ClassTree<T> extends RecursiveElementVisitor<T> {
     _buildSwitches(str, 'constructorName', entries.keys, (name) {
       var constructor = entries[name]!;
       if (verbose) {
-        print(constructor);
+        log.info('[CONSTRUCTOR] $constructor');
       }
 
       var declaringType = constructor.declaringType!.typeNameResolvable;
@@ -1558,7 +1679,7 @@ class _ClassTree<T> extends RecursiveElementVisitor<T> {
       var field = entries[name]!;
 
       if (verbose) {
-        print(field);
+        log.info('[FIELD] $field');
       }
 
       var fieldDeclaringType = field.declaringType;
@@ -1605,7 +1726,7 @@ class _ClassTree<T> extends RecursiveElementVisitor<T> {
     _buildSwitches(str, 'fieldName', entries.keys, (name) {
       var field = entries[name]!;
       if (verbose) {
-        print(field);
+        log.info('[FIELD] $field');
       }
 
       var declaringType = field.declaringType!.typeNameResolvable;
@@ -1671,7 +1792,7 @@ class _ClassTree<T> extends RecursiveElementVisitor<T> {
     _buildSwitches(str, 'methodName', entries.keys, (name) {
       var method = entries[name]!;
       if (verbose) {
-        print(method);
+        log.info('[METHOD] $method');
       }
 
       var declaringType = method.declaringType!.typeNameResolvable;
@@ -1704,7 +1825,7 @@ class _ClassTree<T> extends RecursiveElementVisitor<T> {
     _buildSwitches(str, 'methodName', entries.keys, (name) {
       var method = entries[name]!;
       if (verbose) {
-        print(method);
+        log.info('[METHOD] $method');
       }
 
       var declaringType = method.declaringType!.typeNameResolvable;